dcodeIO/protobuf.js#876
Simple prototype to support use of (custom) type-converters for static code. It should generally reduce file size also.
If there's a desire to really reduce static code size, similar approach could be used for `verify(), encode(), decode()` methods

Note:
 * `cli/targets/static-converter.js` is a copy of the original `cli/targets/static.js` but the converter implementation has been replaced.
 * `.scripts/gentypeconvertertests.js` is a copy of `.scripts/gentests.js` but params have been modified to enable testing of the prototype

This is only a basic prototype, it is not fully functional. It does generate code, run:
`node ./scripts/gentypeconvertertests.js`

I'm looking for feedback, and indications if anyone is interested in such a feature.